refactor: rename vm/ to sandbox/ and add stub packages for remaining components
- Rename internal/vm/ to internal/sandbox/ to match project architecture
- Add stub packages: internal/boot/, internal/image/, internal/virtio/
- Add platform-specific network implementations (vmnet_darwin.go, linux network)
- Update CLI commands to use sandbox package instead of vm package
- Add storage manager interface for snapshot operations
- Update test imports to reflect new package structure
